save all temporary files in the working directory
Now the working directory is placed in a config file stored in the user's
home directory:

~/.ptycho_gui_config

so that the recon code knows where it is. All temporary files are now
written to the working directory.

TODO: repopulate the working directory when opening the GUI if the config
file is found.