can fetch experimental parameters from h5
By clicking "View data frame" now the experimental parameters are loaded
from h5. In the future we will add more functionalities to this button,
but for now this meets the beamline people's need.

TODO: modify the param accordingly if users change those parameters in
the GUI. There could be situations in which users do not want to use the
values loaded from either the databroker or the h5 to do reconstruction.