loose tie to Open MPI; remember config history by default; CUDA installation is not necessary
1. In both the GUI itself and the underlying ptycho code, there were several
places that explictly required Open MPI, which caused enormous issues that are
hard to resolve. Now everything is independent of the underlying MPI implementation
as long as mpi4py is built properly. A machine file for MPICH is provided.

2. Config history (a Param object in text, basically) is by default written
to the user's home directory. This is necessary because the ptycho code needs
to know where the working directory is, and so to avoid a loop the config
cannot be saved there. But perhaps there's a better approach? (TODO) The config
history can also be cleared in the menu bar.

3. if the user just wants to use the CPU version, it's not necessary to install
CUDA and everything that depends on it. This is handled properly in the ptycho
code.