support launching multi-GPU using MPI machine file; bug fix
A rare bug (QThread: Destroyed while thread is still running) happens when,
I think, a reconstruction about to finish is forced to terminate. The fix
is suggested in https://stackoverflow.com/questions/43647719/ but I'm not
sure if it will actually fix the problem since it's hard to reproduce (I
only encountered it 3 times so far).