reduce the chance of not clearing shared memory
Previously if the reconstruction is terminated early (but after the backend
has allocated the shared memory segments), the GUI is unable to properly
clear them out. Now the chance of hitting this is significantly reduced by
using Qt's signal mechanism. Once the shared memory is allocated, the GUI
will immediately get a handle to it.

The only chance of not invoking the destructor is when the program is being
terminated while Qt handles the signal, but during the experiment I wasn't
able to hit this. It seems hard?